@@ -20,16 +20,16 @@ classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
   "Programming Language :: R",
   "Topic :: Scientific/Engineering",
   "Topic :: Scientific/Engineering :: Information Analysis",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
   "click",
   "genno >= 1.20",
@@ -63,10 +63,10 @@ tests = [
   "ixmp[report,tutorial]",
   "memory_profiler",
   "nbclient >= 0.5",
-  "pretenders >= 1.4.4",
   "pytest >= 5",
   "pytest-benchmark",
   "pytest-cov",
+  "pytest-httpserver",
   "pytest-rerunfailures",
   "pytest-xdist",
 ]
@@ -81,12 +81,16 @@ exclude_also = [
 ]
 omit = ["ixmp/util/sphinx_linkcode_github.py"]
 
+[tool.mypy]
+exclude = [
+  "build/",
+]
+
 [[tool.mypy.overrides]]
 # Packages/modules for which no type hints are available.
 module = [
   "jpype",
   "memory_profiler",
-  "pretenders.*",
   "pyam",
 ]
 ignore_missing_imports = true
