@@ -1,8 +1,6 @@
-from typing import Dict, Type
-
 #: Mapping from names to available models. To register additional models,
 #: add elements to this variable.
-MODELS: Dict[str, Type] = {}
+MODELS: dict[str, type] = {}
 
 
 def get_model(name, **model_options):
