@@ -1,5 +1,3 @@
-from typing import Dict
-
 #: Dimensions to rename when extracting raw data from Scenario objects.
 #: Mapping from Scenario dimension name -> preferred dimension name.
-RENAME_DIMS: Dict[str, str] = {}
+RENAME_DIMS: dict[str, str] = {}
