@@ -1,5 +1,4 @@
 import re
-import sys
 from pathlib import Path
 
 import pandas as pd
@@ -422,12 +421,7 @@ def test_solve(ixmp_cli, test_mp):
     ]
     result = ixmp_cli.invoke(cmd)
     assert result.exit_code == 1, result.output
-    exp = (
-        "='non-existing'"
-        if sys.version_info.minor != 12
-        else ", scenario, or version not found"
-    )
-    assert f"Error: model{exp}" in result.output
+    assert "Error: model='non-existing'" in result.output
 
     result = ixmp_cli.invoke([f"--url=ixmp://{test_mp.name}/foo/bar", "solve"])
     assert UsageError.exit_code == result.exit_code, result.output
